refactor: rename folder name #49
